+++ CHANGELOG.md
@@ -16,6 +16,10 @@ For a full diff see [`0.12.0...0.12.1`](https://github.com/localheinz/phpstan-ru
 
 * Started ignoring interfaces from analysis by `Methods\FinalInAbstractClassRule` to avoid inappropriate errors ([#132](https://github.com/localheinz/phpstan-rules/pull/132)), by [@localheinz](https://github.com/localheinz)
 
+### Added
+
+* Added `Methods\FinalInAbstractClassRule`, which reports an error when a method in a `final` class is `protected` but could be `private` ([#125](https://github.com/localheinz/phpstan-rules/pull/125)), by [@localheinz](https://github.com/localheinz)
+
 ## [`0.12.1`](https://github.com/localheinz/phpstan-rules/releases/tag/0.12.1)
 
 For a full diff see [`0.12.0...0.12.1`](https://github.com/localheinz/phpstan-rules/compare/0.12.0...0.12.1).